requirements.txt
django-filebrowser==3.12.1
